saturn/requirements.txt
fastapi[standard]
chromadb
taskiq
taskiq[reload]
taskiq-redis
taskiq-fastapi
redis
numpy==1.25.2
pandas==2.2.2
fuzzywuzzy==0.18.0
python-Levenshtein
opencv-python==4.10.0.84
moviepy==1.0.3
pydub==0.25.1
openai-whisper==20230124
ultralytics==8.2.48
pytesseract==0.3.13
Pillow==10.4.0
scikit-learn==1.4.2
librosa==0.10.2.post1
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
